test: unify MetamodelTest and MetaModelTest
